+++ src/ccp_pred1.c
@@ -503,31 +503,33 @@ Pred1Stat(Context ctx, int dir)
     switch (dir) {
 	case COMP_DIR_XMIT:
 	    Printf("\tBytes\t: %llu -> %llu (%+lld%%)\r\n",
-		p->xmit_stats.InOctets,
-		p->xmit_stats.OutOctets,
+		(unsigned long long)p->xmit_stats.InOctets,
+		(unsigned long long)p->xmit_stats.OutOctets,
 		((p->xmit_stats.InOctets!=0)?
-		    ((int64_t)(p->xmit_stats.OutOctets - p->xmit_stats.InOctets)*100/(int64_t)p->xmit_stats.InOctets):
+		    ((unsigned long long)(p->xmit_stats.OutOctets - p->xmit_stats.InOctets)
+			*100/(unsigned long long)p->xmit_stats.InOctets):
 		    0));
 	    Printf("\tFrames\t: %llu -> %lluc + %lluu\r\n",
-		p->xmit_stats.FramesPlain,
-		p->xmit_stats.FramesComp,
-		p->xmit_stats.FramesUncomp);
+		(unsigned long long)p->xmit_stats.FramesPlain,
+		(unsigned long long)p->xmit_stats.FramesComp,
+		(unsigned long long)p->xmit_stats.FramesUncomp);
 	    Printf("\tErrors\t: %llu\r\n",
-		p->recv_stats.Errors);
+		(unsigned long long)p->recv_stats.Errors);
 	    break;
 	case COMP_DIR_RECV:
 	    Printf("\tBytes\t: %llu <- %llu (%+lld%%)\r\n",
-		p->recv_stats.OutOctets,
-		p->recv_stats.InOctets,
+		(unsigned long long)p->recv_stats.OutOctets,
+		(unsigned long long)p->recv_stats.InOctets,
 		((p->recv_stats.OutOctets!=0)?
-		    ((int64_t)(p->recv_stats.InOctets - p->recv_stats.OutOctets)*100/(int64_t)p->recv_stats.OutOctets):
+		    ((unsigned long long)(p->recv_stats.InOctets - p->recv_stats.OutOctets)
+			*100/(unsigned long long)p->recv_stats.OutOctets):
 		    0));
 	    Printf("\tFrames\t: %llu <- %lluc + %lluu\r\n",
-		p->xmit_stats.FramesPlain,
-		p->xmit_stats.FramesComp,
-		p->xmit_stats.FramesUncomp);
+		(unsigned long long)p->xmit_stats.FramesPlain,
+		(unsigned long long)p->xmit_stats.FramesComp,
+		(unsigned long long)p->xmit_stats.FramesUncomp);
 	    Printf("\tErrors\t: %llu\r\n",
-		p->recv_stats.Errors);
+		(unsigned long long)p->recv_stats.Errors);
     	    break;
 	default:
     	    assert(0);
